Re-use code for wrapping/unwrapping return types
